local.requirements.txt
'''
We recommend running the local.requirements.txt with psycopg2 omitted only if you just 
need the dependencies to set up your virtual environment locally. If you want to run the 
project locally, you will need to set up PostgreSQL as shown in the RUN process in our 
Dockerfile.
'''
asgiref==3.8.1
Django==3.2.25
django-filter==23.5
djangorestframework==3.15.1
importlib_metadata==7.1.0
Markdown==3.6
#psycopg2==2.8.6
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.12.1
zipp==3.19.1
django-injector==0.3.1
